+++ core/src/main/java/com/ericsson/bss/cassandra/ecchronos/core/metrics/TableRepairMetrics.java
@@ -59,14 +59,14 @@ public interface TableRepairMetrics
     void repairTiming(TableReference tableReference, long timeTaken, TimeUnit timeUnit, boolean successful);
 
     /**
-     * Report failed repair attempt.
+     * Report that repair task has failed.
      *
      * @param tableReference The table the repair was performed on.
      */
     void failedRepairTask(TableReference tableReference);
 
     /**
-     * Report succeeded repair attempt.
+     * Report that repair task has succeeded.
      *
      * @param tableReference The table the repair was performed on.
      */