+++ CHANGES.md
@@ -2,7 +2,7 @@
 
 ## Version 4.0.0
 
-* Add metric for failed repair sessions - Issue #295
+* Add metric for failed and succeeded repair tasks - Issue #295
 * Migrate to datastax driver-4.14.1 - Issue #269
 
 ## Version 3.0.0